FIx - Unified UserSearchResult type that serves both the friends and users features, as well as any other components that need user search results.
