+++ Web/src/features/users/index.ts
@@ -32,4 +32,4 @@ export { UserSearch } from './components/UserSearch';
 export { UserSearchResult } from './components/UserSearchResult';
 
 // Re-export commonly used types for convenience
-export type { UserSearchResult as User } from './types';
+export type { UserSearchResult as User } from '@/types/user';
